+++ core/src/sys_events.rs
@@ -2,32 +2,7 @@ use core::fmt;
 
 // SYSTEM EVENTS
 // ================================================================================================
-
-// Randomly generated constant values for the VM's system events. All values were sampled
-// between 0 and 2^32.
-pub use constants::*;
-
 use crate::EventId;
-
-#[rustfmt::skip]
-mod constants {
-    pub const EVENT_MERKLE_NODE_MERGE: u32            = 276124218;
-    pub const EVENT_MERKLE_NODE_TO_STACK: u32         = 361943238;
-    pub const EVENT_MAP_VALUE_TO_STACK: u32           = 574478993;
-    pub const EVENT_MAP_VALUE_TO_STACK_N: u32         = 630847990;
-    pub const EVENT_HAS_MAP_KEY: u32                  = 652777600;
-    pub const EVENT_EXT2_INV: u32                     = 1251967401;
-    pub const EVENT_U32_CLZ: u32                      = 1951932030;
-    pub const EVENT_U32_CTZ: u32                      = 2008979519;
-    pub const EVENT_U32_CLO: u32                      = 2032895094;
-    pub const EVENT_U32_CTO: u32                      = 2083700134;
-    pub const EVENT_ILOG2: u32                        = 2297972669;
-    pub const EVENT_MEM_TO_MAP: u32                   = 2389394361;
-    pub const EVENT_HDWORD_TO_MAP: u32                = 2391452729;
-    pub const EVENT_HDWORD_TO_MAP_WITH_DOMAIN: u32    = 2822590340;
-    pub const EVENT_HQWORD_TO_MAP: u32                = 2913039991;
-    pub const EVENT_HPERM_TO_MAP: u32                 = 3297060969;
-}
 
 /// Defines a set of actions which can be initiated from the VM to inject new data into the advice
 /// provider.
@@ -38,6 +13,7 @@ mod constants {
 /// All actions, except for `MerkleNodeMerge`, `Ext2Inv` and `UpdateMerkleNode` can be invoked
 /// directly from Miden assembly via dedicated instructions.
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
+#[repr(u32)]
 pub enum SystemEvent {
     // MERKLE STORE EVENTS
     // --------------------------------------------------------------------------------------------
@@ -261,50 +237,35 @@ pub enum SystemEvent {
     HpermToMap,
 }
 
-impl SystemEvent {
-    pub fn into_event_id(self) -> u32 {
-        match self {
-            SystemEvent::MerkleNodeMerge => EVENT_MERKLE_NODE_MERGE,
-            SystemEvent::MerkleNodeToStack => EVENT_MERKLE_NODE_TO_STACK,
-            SystemEvent::MapValueToStack => EVENT_MAP_VALUE_TO_STACK,
-            SystemEvent::MapValueToStackN => EVENT_MAP_VALUE_TO_STACK_N,
-            SystemEvent::HasMapKey => EVENT_HAS_MAP_KEY,
-            SystemEvent::Ext2Inv => EVENT_EXT2_INV,
-            SystemEvent::U32Clz => EVENT_U32_CLZ,
-            SystemEvent::U32Ctz => EVENT_U32_CTZ,
-            SystemEvent::U32Clo => EVENT_U32_CLO,
-            SystemEvent::U32Cto => EVENT_U32_CTO,
-            SystemEvent::ILog2 => EVENT_ILOG2,
-            SystemEvent::MemToMap => EVENT_MEM_TO_MAP,
-            SystemEvent::HdwordToMap => EVENT_HDWORD_TO_MAP,
-            SystemEvent::HdwordToMapWithDomain => EVENT_HDWORD_TO_MAP_WITH_DOMAIN,
-            SystemEvent::HqwordToMap => EVENT_HQWORD_TO_MAP,
-            SystemEvent::HpermToMap => EVENT_HPERM_TO_MAP,
-        }
+impl From<SystemEvent> for EventId {
+    fn from(system_event: SystemEvent) -> Self {
+        EventId::from_u64(system_event as u32 as u64)
     }
+}
 
-    /// Returns a system event corresponding to the specified event ID, or `None` if the event
-    /// ID is not recognized.
-    pub fn from_event_id(event_id: EventId) -> Option<Self> {
-        let event_id: u32 = event_id.as_felt().as_int().try_into().ok()?;
+impl TryFrom<EventId> for SystemEvent {
+    type Error = u64;
+
+    fn try_from(event_id: EventId) -> Result<Self, Self::Error> {
+        let event_id = event_id.as_felt().as_int();
         match event_id {
-            EVENT_MERKLE_NODE_MERGE => Some(SystemEvent::MerkleNodeMerge),
-            EVENT_MERKLE_NODE_TO_STACK => Some(SystemEvent::MerkleNodeToStack),
-            EVENT_MAP_VALUE_TO_STACK => Some(SystemEvent::MapValueToStack),
-            EVENT_MAP_VALUE_TO_STACK_N => Some(SystemEvent::MapValueToStackN),
-            EVENT_HAS_MAP_KEY => Some(SystemEvent::HasMapKey),
-            EVENT_EXT2_INV => Some(SystemEvent::Ext2Inv),
-            EVENT_U32_CLZ => Some(SystemEvent::U32Clz),
-            EVENT_U32_CTZ => Some(SystemEvent::U32Ctz),
-            EVENT_U32_CLO => Some(SystemEvent::U32Clo),
-            EVENT_U32_CTO => Some(SystemEvent::U32Cto),
-            EVENT_ILOG2 => Some(SystemEvent::ILog2),
-            EVENT_MEM_TO_MAP => Some(SystemEvent::MemToMap),
-            EVENT_HDWORD_TO_MAP => Some(SystemEvent::HdwordToMap),
-            EVENT_HDWORD_TO_MAP_WITH_DOMAIN => Some(SystemEvent::HdwordToMapWithDomain),
-            EVENT_HQWORD_TO_MAP => Some(SystemEvent::HqwordToMap),
-            EVENT_HPERM_TO_MAP => Some(SystemEvent::HpermToMap),
-            _ => None,
+            0 => Ok(SystemEvent::MerkleNodeMerge),
+            1 => Ok(SystemEvent::MerkleNodeToStack),
+            2 => Ok(SystemEvent::MapValueToStack),
+            3 => Ok(SystemEvent::MapValueToStackN),
+            4 => Ok(SystemEvent::HasMapKey),
+            5 => Ok(SystemEvent::Ext2Inv),
+            6 => Ok(SystemEvent::U32Clz),
+            7 => Ok(SystemEvent::U32Ctz),
+            8 => Ok(SystemEvent::U32Clo),
+            9 => Ok(SystemEvent::U32Cto),
+            10 => Ok(SystemEvent::ILog2),
+            11 => Ok(SystemEvent::MemToMap),
+            12 => Ok(SystemEvent::HdwordToMap),
+            13 => Ok(SystemEvent::HdwordToMapWithDomain),
+            14 => Ok(SystemEvent::HqwordToMap),
+            15 => Ok(SystemEvent::HpermToMap),
+            other => Err(other),
         }
     }
 }